add dataloaders for downstream task
