+++ BERT_retraining/DataSetup/Dataloaders.py
@@ -5,21 +5,17 @@ sys.path.append(".")
 from torch.utils import data
 import numpy as np
 
-class PretrainingDataset(data.Dataset):
+
+class DownstreamDataset(data.Dataset):
     def __init__(self, input_dict):
         self.input_dict = input_dict
 
     def __len__(self):
-        return len(self.input_dict["input_ids"])
+        return len(self.input_dict["labels"])
 
     def __getitem__(self, index):
-        input_ids = np.array(self.input_dict["input_ids"][index])
-        input_mask = np.array(self.input_dict["input_mask"][index])
-        segment_ids = np.array(self.input_dict["segment_ids"][index])
-        masked_lm_positions = np.array(self.input_dict["masked_lm_positions"][index])
-        masked_lm_ids = np.array(self.input_dict["masked_lm_ids"][index])
-        masked_lm_weights = np.array(self.input_dict["masked_lm_weights"][index])
-        next_sentence_labels = np.array(self.input_dict["next_sentence_labels"][index])
-
-        return input_ids, input_mask, segment_ids, masked_lm_positions, masked_lm_ids,\
-               masked_lm_weights, next_sentence_labels
+        ans_a = np.array(self.input_dict["ans_a"][index])
+        ans_b = np.array(self.input_dict["ans_a"][index])
+        ans_c = np.array(self.input_dict["ans_a"][index])
+        label = np.array(self.input_dict['label'][index])
+        return ans_a, ans_b, ans_c, label
