+++ Downstream/Preprocess/Preprocess.py
@@ -4,7 +4,7 @@ sys.path.append(".")
 import pandas as pd
 from transformers import DistilBertTokenizer
 from tqdm import tqdm
-from utils import save_dictionary
+from Downstream.utils import save_dictionary
 
 
 class Preprocessor:
@@ -66,6 +66,16 @@ class Preprocessor:
                 input_line = answer + context + question
                 self.input_dict[key][ans].append(input_line + [0 for _ in range(512 - len(input_line))])
             self.input_dict[key]["label"].append(labels[i])
+
+    def get_loaders(self, load_flag=False):
+        train_features, valid_features = self.get_features(load_flag=load_flag)
+        train_dataset = PretrainingDataset(input_dict=train_features)
+        valid_dataset = PretrainingDataset(input_dict=valid_features)
+
+        loader_args = dict(shuffle=True, batch_size=con.BATCH_SIZE)
+
+        self.train_loaders = data.DataLoader(train_dataset, **loader_args)
+        self.valid_loaders = data.DataLoader(valid_dataset, **loader_args)
 
     def setup(self):
         self.load_data()
